Give every typing case file exactly one callable
The clip and slice filters, the self-typed filters, `connectivity` and its
two callers, `add_text` and `add_title`, and the `__getitem__`/`__setitem__`
pairs each get a file named after the callable, with the setup each needs.
The contributor guide states the naming rule.
