+++ tests/typing/cases/find_containing_cell.py
@@ -14,7 +14,7 @@ def an_image() -> pv.ImageData:
 
 
 SKIP_RUNTIME = {
-    'an_image().find_containing_cell([(1.0, 1.0, 1.0), (0.5, 0.5, 0.5)])': ('the runtime checker has no `npt_promote`, so an int64 array is not a `NumpyArray[int]`'),
+    'an_image().find_containing_cell([(1.0, 1.0, 1.0), (0.5, 0.5, 0.5)])': ('the runtime checker does not accept an int64 array as `NumpyArray[int]`'),
 }
